requirements.txt
Django==3.2.9
dr-scaffold==2.1.1
djangorestframework==3.12.4
django-seed==0.3.1
python-decouple==3.5
django-extensions==3.1.5
psycopg2-binary==2.9.2
